Refactored the code to move reusable method to module utils
